+++ packages/client/CHANGELOG.md
@@ -9,6 +9,8 @@
   - Export `WhitelistEntriesByChain` type from descriptors to help build a chain-based whitelist.
 - **Codegen**:
   - Whitelist accepts an object of `chain key` -> `chain whitelist`.
+- **PolkadotSDKCompat**
+  - Add `fixMissingInitialBest` enhancer which addresses [the following issue](https://github.com/paritytech/polkadot-sdk/pull/10525) on the PolkadotSDK node.
 
 ### Fixed