+++ packages/client/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.23.0 - 2025-12-08
+
 ### Added
 
 - **CLI**:
@@ -19,6 +21,10 @@
   - Errors not being properly identified when substrate-client is nested in other packages.
 - **Codegen**:
   - WhitelistEntries type helper is effected by the whitelist setting.
+- **Smoldot:**
+  - Update smoldot to `2.0.40`
+- **Chains:**:
+  - Update `lightSyncState`
 
 ## 1.22.0 - 2025-11-27